+++ thesis/Models/totalNoFitForHumanConsumptions.cs
@@ -10,7 +10,7 @@ namespace thesis.Models
         public int Id { get; set; }
         public Species Species { get; set; }
         public int NoOfHeads { get; set; }
-        public int DressedWeight { get; set; }
+        public double DressedWeight { get; set; }
         [ForeignKey("PostmortemId")]
         public int PostmortemId { get; set; }
         public Postmortem? Postmortem { get; set; }
